Improve landing page SEO and footer
