+++ src/styles/globals.css
@@ -354,3 +354,59 @@
     }
   }
        
+/* Landing Page Styles */
+.landing-wrapper {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding: 60px 16px;
+  min-height: 100vh;
+}
+
+.landing-hero {
+  text-align: center;
+  margin-bottom: 48px;
+}
+
+.cta-button {
+  display: inline-block;
+  margin-top: 24px;
+  padding: 12px 24px;
+  background-color: #22c55e;
+  color: #000;
+  font-weight: 600;
+  border-radius: 8px;
+  text-decoration: none;
+  transition: background 0.2s ease;
+}
+
+.cta-button:hover {
+  background-color: #16a34a;
+}
+
+.features {
+  display: flex;
+  flex-direction: column;
+  gap: 32px;
+  max-width: 600px;
+}
+
+.feature {
+  background: #181818;
+  padding: 24px;
+  border-radius: 12px;
+  border-left: 3px solid #22c55e;
+}
+
+.feature h3 {
+  margin: 0 0 8px 0;
+  font-size: 1.2rem;
+  color: #ffffff;
+}
+
+.feature p {
+  margin: 0;
+  color: #cccccc;
+  font-size: 0.95rem;
+  line-height: 1.6;
+}
